ocr_service/requirements.txt
flask
Pillow
pytesseract